refactor: on-chain-data -> on-chain-data-client
