+++ src/on-chain-data-client/hotels/hotel.js
