+++ src/on-chain-data-client/hotels/wt-index.js
