+++ src/on-chain-data-client/storage-pointer.js
