+++ test/on-chain-data-client/hotels/hotel.spec.js
@@ -1,9 +1,9 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
 import helpers from '../../utils/helpers';
-import OnChainAirline from '../../../src/on-chain-data/airlines/airline';
+import OnChainHotel from '../../../src/on-chain-data-client/hotels/hotel';
 
-describe('WTLibs.on-chain-data.airlines.Airline', () => {
+describe('WTLibs.on-chain-data.hotels.Hotel', () => {
   let contractsStub, createdStub, utilsStub, indexContractStub, urlStub, managerStub;
   const validUri = 'schema://new-url';
   const validManager = 'manager';
@@ -18,7 +18,7 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
     managerStub = helpers.stubContractMethodResult('some-remote-manager');
     createdStub = helpers.stubContractMethodResult('created-block');
     contractsStub = {
-      getAirlineInstance: sinon.stub().resolves({
+      getHotelInstance: sinon.stub().resolves({
         methods: {
           dataUri: urlStub,
           manager: managerStub,
@@ -27,7 +27,7 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
         },
       }),
       decodeLogs: sinon.stub().returns([{
-        attributes: [{ value: '0xnew-airline-address' }],
+        attributes: [{ value: '0xnew-hotel-address' }],
       }]),
     };
     indexContractStub = {
@@ -35,17 +35,17 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
         address: 'index-address',
       },
       methods: {
-        callAirline: helpers.stubContractMethodResult('called-airline'),
-        registerAirline: helpers.stubContractMethodResult('registered-airline'),
-        deleteAirline: helpers.stubContractMethodResult('deleted-airline'),
-        transferAirline: helpers.stubContractMethodResult('transfer-airline'),
+        callHotel: helpers.stubContractMethodResult('called-hotel'),
+        registerHotel: helpers.stubContractMethodResult('registered-hotel'),
+        deleteHotel: helpers.stubContractMethodResult('deleted-hotel'),
+        transferHotel: helpers.stubContractMethodResult('transfer-hotel'),
       },
     };
   });
 
   describe('toPlainObject', () => {
     it('should return a plain JS object', async () => {
-      const provider = OnChainAirline.createInstance(utilsStub, contractsStub, indexContractStub);
+      const provider = OnChainHotel.createInstance(utilsStub, contractsStub, indexContractStub);
       await provider.setLocalData({ dataUri: validUri, manager: validManager });
       // initialize dataIndex so we're able to mock it later
       await provider.dataIndex;
@@ -58,22 +58,22 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
           },
         },
       });
-      const plainAirline = await provider.toPlainObject(); // fields?
-      assert.equal(plainAirline.manager, validManager);
-      assert.isUndefined(plainAirline.toPlainObject);
-      assert.equal(plainAirline.dataUri.ref, validUri);
-      assert.isDefined(plainAirline.dataUri.contents);
-      assert.isDefined(plainAirline.dataUri.contents.descriptionUri);
+      const plainHotel = await provider.toPlainObject(); // fields?
+      assert.equal(plainHotel.manager, validManager);
+      assert.isUndefined(plainHotel.toPlainObject);
+      assert.equal(plainHotel.dataUri.ref, validUri);
+      assert.isDefined(plainHotel.dataUri.contents);
+      assert.isDefined(plainHotel.dataUri.contents.descriptionUri);
     });
   });
 
   describe('createOnChainData', () => {
     it('should return transaction metadata', async () => {
-      const provider = OnChainAirline.createInstance(utilsStub, contractsStub, indexContractStub);
+      const provider = OnChainHotel.createInstance(utilsStub, contractsStub, indexContractStub);
       const callSpy = sinon.spy(provider, '_createOnChainData');
       const result = await provider.createOnChainData({ from: 'xx' });
       assert.isDefined(result.transactionData);
-      assert.isDefined(result.airline);
+      assert.isDefined(result.hotel);
       assert.isUndefined(result.record);
       assert.isDefined(result.eventCallbacks);
       assert.isDefined(result.eventCallbacks.onReceipt);
@@ -84,13 +84,13 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
 
   describe('updateOnChainData', () => {
     it('should return transaction metadata', async () => {
-      const provider = OnChainAirline.createInstance(utilsStub, contractsStub, indexContractStub, 'fake-address');
+      const provider = OnChainHotel.createInstance(utilsStub, contractsStub, indexContractStub, 'fake-address');
       const callSpy = sinon.spy(provider, '_updateOnChainData');
       await provider.setLocalData({ dataUri: 'in-memory://new-link' });
       const result = await provider.updateOnChainData({ from: 'xx' });
       assert.equal(result.length, 1);
       assert.isDefined(result[0].transactionData);
-      assert.isDefined(result[0].airline);
+      assert.isDefined(result[0].hotel);
       assert.isUndefined(result[0].record);
       assert.isDefined(result[0].eventCallbacks);
       assert.isDefined(result[0].eventCallbacks.onReceipt);
@@ -101,11 +101,11 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
 
   describe('transferOnChainOwnership', () => {
     it('should return transaction metadata', async () => {
-      const provider = OnChainAirline.createInstance(utilsStub, contractsStub, indexContractStub, 'fake-address');
+      const provider = OnChainHotel.createInstance(utilsStub, contractsStub, indexContractStub, 'fake-address');
       const callSpy = sinon.spy(provider, '_transferOnChainOwnership');
       const result = await provider.transferOnChainOwnership('new-manager', { from: 'xx' });
       assert.isDefined(result.transactionData);
-      assert.isDefined(result.airline);
+      assert.isDefined(result.hotel);
       assert.isUndefined(result.record);
       assert.isDefined(result.eventCallbacks);
       assert.isDefined(result.eventCallbacks.onReceipt);
@@ -116,11 +116,11 @@ describe('WTLibs.on-chain-data.airlines.Airline', () => {
 
   describe('removeOnChainData', () => {
     it('should return transaction metadata', async () => {
-      const provider = OnChainAirline.createInstance(utilsStub, contractsStub, indexContractStub, 'fake-address');
+      const provider = OnChainHotel.createInstance(utilsStub, contractsStub, indexContractStub, 'fake-address');
       const callSpy = sinon.spy(provider, '_removeOnChainData');
       const result = await provider.removeOnChainData({ from: 'xx' });
       assert.isDefined(result.transactionData);
-      assert.isDefined(result.airline);
+      assert.isDefined(result.hotel);
       assert.isUndefined(result.record);
       assert.isDefined(result.eventCallbacks);
       assert.isDefined(result.eventCallbacks.onReceipt);
