+++ test/on-chain-data-client/index.spec.js
@@ -1,7 +1,7 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
 import { AbstractDataModel, AirlineDataModel } from '../../src/on-chain-data';
-import WTAirlineIndex from '../../src/on-chain-data/airlines/wt-index';
+import WTAirlineIndex from '../../src/on-chain-data-client/airlines/wt-index';
 import testedDataModel from '../utils/data-airline-model-definition';
 
 describe('WTLibs.on-chain-data.DataModel', () => {
