+++ test/on-chain-data-client/remotely-backed-dataset.spec.js
@@ -1,8 +1,8 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import RemotelyBackedDataset from '../../src/on-chain-data/remotely-backed-dataset';
+import RemotelyBackedDataset from '../../src/on-chain-data-client/remotely-backed-dataset';
 
-import { RemoteDataAccessError, RemoteDataReadError } from '../../src/on-chain-data/errors';
+import { RemoteDataAccessError, RemoteDataReadError } from '../../src/on-chain-data-client/errors';
 
 describe('WTLibs.RemotelyBackedDataset', () => {
   let bbbInstance, remoteGetterStub, remoteSetterStub,
